fix: cast Ctim fields to int64 for 32-bit ARM Linux builds

+++ internal/persis/filedoc/birthtime_linux.go
@@ -13,7 +13,7 @@ import (
 // is not available via the standard syscall package. Ctime is used as an approximation.
 func fileCreationTime(info os.FileInfo) time.Time {
 	if stat, ok := info.Sys().(*syscall.Stat_t); ok {
-		return time.Unix(stat.Ctim.Sec, stat.Ctim.Nsec)
+		return time.Unix(int64(stat.Ctim.Sec), int64(stat.Ctim.Nsec))
 	}
 	return info.ModTime()
 }
